restructuring of swiftML algorthms to respective class directory
